refactor(lsposed): move startup/ and statistics/ into their own packages
Two of the cheap leaves from the flat 96-file package: nothing else in the
module reaches into them beyond a handful of entry points, so the move is
mechanical and the tree gains two names that say what they hold.

- startup/: MainActivity, StartupCoordinator, StartupTrace and the
  blocking screen. The manifest and scripts/measure-startup.py follow the
  activity to `.startup.MainActivity` — verified against the built APK,
  which carries dev/okhsunrog/vpnhide/startup/MainActivity
- statistics/: the screen, its cache and the pure counter model

No logic touched; the diff is package lines and imports.
