+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/AgentControl.kt
@@ -3,6 +3,10 @@ package dev.okhsunrog.vpnhide
 import android.content.Context
 import android.net.ConnectivityManager
 import dev.okhsunrog.vpnhide.generated.HookIds
+import dev.okhsunrog.vpnhide.statistics.BackendStatistics
+import dev.okhsunrog.vpnhide.statistics.StatisticsRow
+import dev.okhsunrog.vpnhide.statistics.StatisticsState
+import dev.okhsunrog.vpnhide.statistics.buildStatisticsState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import java.io.File
