+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/DebugExport.kt
@@ -3,6 +3,8 @@ package dev.okhsunrog.vpnhide
 import android.content.Context
 import android.net.ConnectivityManager
 import android.os.Build
+import dev.okhsunrog.vpnhide.statistics.buildStatisticsState
+import dev.okhsunrog.vpnhide.ui.components.container
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
