+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/HookRegistry.kt
@@ -1,6 +1,7 @@
 package dev.okhsunrog.vpnhide
 
 import dev.okhsunrog.vpnhide.generated.HookIds
+import dev.okhsunrog.vpnhide.statistics.parseProtocolStatusBlock
 
 // The set-shaped view over the generated hook registry.
 //
