+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/LegacyImportUi.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import dev.okhsunrog.vpnhide.settings.AppSettings
 import dev.okhsunrog.vpnhide.ui.components.ButtonSpinner
 import dev.okhsunrog.vpnhide.ui.components.EnhancedOutlinedButton
 import kotlinx.coroutines.launch
