+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/LogTags.kt
@@ -1,5 +1,7 @@
 package dev.okhsunrog.vpnhide
 
+import dev.okhsunrog.vpnhide.hook.HookLog
+
 /**
  * Single registry of the logcat tags this project emits, so call sites, the
  * [HookLog] system_server sink and the debug-capture logcat filter
