+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/StateCache.kt
@@ -1,5 +1,6 @@
 package dev.okhsunrog.vpnhide
 
+import dev.okhsunrog.vpnhide.startup.StartupTrace
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
