+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/startup/MainActivity.kt
@@ -1,4 +1,4 @@
-package dev.okhsunrog.vpnhide
+package dev.okhsunrog.vpnhide.startup
 
 import android.Manifest
 import android.os.Bundle
@@ -44,11 +44,36 @@ import androidx.compose.ui.unit.lerp
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
+import dev.okhsunrog.vpnhide.AgentControlBridge
+import dev.okhsunrog.vpnhide.AppListCache
+import dev.okhsunrog.vpnhide.AppSearchTopBar
+import dev.okhsunrog.vpnhide.BackgroundUpdateChecks
+import dev.okhsunrog.vpnhide.DashboardCache
+import dev.okhsunrog.vpnhide.DashboardLoadingState
+import dev.okhsunrog.vpnhide.DashboardScreen
+import dev.okhsunrog.vpnhide.DiagnosticsSettingsScreen
+import dev.okhsunrog.vpnhide.GroundTruthProbe
+import dev.okhsunrog.vpnhide.ProtectionScreen
+import dev.okhsunrog.vpnhide.R
+import dev.okhsunrog.vpnhide.RootSnapshotCache
+import dev.okhsunrog.vpnhide.RoutingGateCache
+import dev.okhsunrog.vpnhide.SelfTargetFailureKind
+import dev.okhsunrog.vpnhide.SettingsScreen
+import dev.okhsunrog.vpnhide.TargetFilterChips
+import dev.okhsunrog.vpnhide.TargetListSortMode
+import dev.okhsunrog.vpnhide.TargetsCache
+import dev.okhsunrog.vpnhide.VpnHideLog
+import dev.okhsunrog.vpnhide.VpnTransportWatcher
+import dev.okhsunrog.vpnhide.isMainAppProfile
 import dev.okhsunrog.vpnhide.settings.AppSettings
 import dev.okhsunrog.vpnhide.settings.LocalSettingsInteractor
 import dev.okhsunrog.vpnhide.settings.LocalSettingsState
 import dev.okhsunrog.vpnhide.settings.RepositorySettingsInteractor
 import dev.okhsunrog.vpnhide.settings.SettingsRepository
+import dev.okhsunrog.vpnhide.shouldRequestUpdateNotificationPermission
+import dev.okhsunrog.vpnhide.statistics.StatisticsCache
+import dev.okhsunrog.vpnhide.statistics.StatisticsScreen
+import dev.okhsunrog.vpnhide.suExec
 import dev.okhsunrog.vpnhide.ui.components.BlockingErrorCard
 import dev.okhsunrog.vpnhide.ui.components.ButtonSpinner
 import dev.okhsunrog.vpnhide.ui.components.EnhancedButton
