+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/startup/StartupCoordinator.kt
@@ -1,6 +1,32 @@
-package dev.okhsunrog.vpnhide
+package dev.okhsunrog.vpnhide.startup
 
 import android.content.Context
+import dev.okhsunrog.vpnhide.AppAutoHideSignal
+import dev.okhsunrog.vpnhide.AppListCache
+import dev.okhsunrog.vpnhide.BuildConfig
+import dev.okhsunrog.vpnhide.CanonicalConfig
+import dev.okhsunrog.vpnhide.CanonicalConfigRepository
+import dev.okhsunrog.vpnhide.DashboardCache
+import dev.okhsunrog.vpnhide.DashboardState
+import dev.okhsunrog.vpnhide.DiagnosticsCache
+import dev.okhsunrog.vpnhide.LogTags
+import dev.okhsunrog.vpnhide.PackageInventorySeed
+import dev.okhsunrog.vpnhide.RootSnapshot
+import dev.okhsunrog.vpnhide.RootSnapshotCache
+import dev.okhsunrog.vpnhide.RoutingGateCache
+import dev.okhsunrog.vpnhide.SelfTargetFailureKind
+import dev.okhsunrog.vpnhide.SelfTargetPreparation
+import dev.okhsunrog.vpnhide.TargetsCache
+import dev.okhsunrog.vpnhide.UpdateCheckCache
+import dev.okhsunrog.vpnhide.VpnHideLog
+import dev.okhsunrog.vpnhide.canonicalConfigForStartupDebugReconcile
+import dev.okhsunrog.vpnhide.cleanupStaleZygiskStatus
+import dev.okhsunrog.vpnhide.ensureSelfInTargets
+import dev.okhsunrog.vpnhide.next
+import dev.okhsunrog.vpnhide.parseTargetsSnapshot
+import dev.okhsunrog.vpnhide.reconcileAutoHiddenPackages
+import dev.okhsunrog.vpnhide.runRuntimeConfigReconcile
+import dev.okhsunrog.vpnhide.toAutoHideSignal
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
