+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/statistics/StatisticsCache.kt
@@ -1,5 +1,8 @@
-package dev.okhsunrog.vpnhide
+package dev.okhsunrog.vpnhide.statistics
 
+import dev.okhsunrog.vpnhide.LogTags
+import dev.okhsunrog.vpnhide.RootSnapshotCache
+import dev.okhsunrog.vpnhide.StateCache
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.StateFlow
