+++ lsposed/app/src/main/kotlin/dev/okhsunrog/vpnhide/ui/components/BlockingErrorCard.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import dev.okhsunrog.vpnhide.next
 
 /**
  * The shared "can't proceed" card. Every full-screen blocking state in the app
